pyproject.toml
[project]
name = "fastapiproject1"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi-pagination>=0.13.1",
    "fastapi>=0.115.12",
    "pydantic-settings>=2.9.1",
    "pymysql>=1.1.1",
    "sqlalchemy>=2.0.40",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.2",
    "watchdog>=6.0.0",
]
